pyproject.toml
[project]
name = "5-day-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.86.0",
    "python-dotenv>=1.0.0",
]
